Added authenticate

+++ VnStyle/VnStyle/VnStyle.Services/Data/Domain/Memberships/AspNetUserRoles.cs
@@ -2,7 +2,7 @@ namespace VnStyle.Services.Data.Domain.Memberships
 {
     public class AspNetUserRole 
     {
-        public long UserId { get; set; }
-        public long RoleId { get; set; }
+        public int UserId { get; set; }
+        public int RoleId { get; set; }
     }
 }
